requirements.txt
Flask
docx-mailmerge
docx-mailmerge